% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiKernel.R
\name{cvMultiKernel}
\alias{cvMultiKernel}
\title{Cross-validation for multivariate kernel regression}
\usage{
cvMultiKernel(response, covariate, confounder = NULL, kernel = c("linear",
  "quadratic", "gaussian"), intercept = TRUE, tau, K = 5, pure = FALSE)
}
\arguments{
\item{response}{matrix of response variables}

\item{covariate}{matrix of covariate variables, which are included in the kernel.}

\item{confounder}{matrix or data.frame of confounder variables, which are not included in the kernel.}

\item{kernel}{Type of kernel to use.}

\item{tau}{Tuning parameter.}

\item{K}{number of folds for cross-validation.}

\item{pure}{Logical. Use the pure R version?}
}
\value{
Returns the estimated prediction error, as well as a separate value for each fold
}
\description{
This function performs cross-validation for multivariate kernel regression
}

