% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiKernel.R
\name{fitMultiKernel}
\alias{fitMultiKernel}
\title{Multivariate kernel regression}
\usage{
fitMultiKernel(response, covariate, confounder = NULL, kernel = c("linear",
  "quadratic", "gaussian"), tau, pure = FALSE)
}
\arguments{
\item{response}{matrix of response variables}

\item{covariate}{matrix of covariate variables, which are included in the kernel.}

\item{confounder}{matrix or data.frame of confounder variables, which are not included in the kernel.}

\item{kernel}{Type of kernel to use.}

\item{tau}{Tuning parameter.}

\item{pure}{Logical. Use the pure R version?}
}
\value{
Returns the kernel predictor.
}
\description{
This function performs multivariate kernel regression by minimizing a specific loss function
}

