% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelFunctions.R
\name{linearKernel}
\alias{IBSkernel}
\alias{gaussKernel}
\alias{linearKernel}
\alias{quadraticKernel}
\title{Kernel functions}
\usage{
linearKernel(X, X_test = X)

quadraticKernel(X, X_test = X)

IBSkernel(X, X_test = X)

gaussKernel(X, X_test = X, rho = 1)
}
\arguments{
\item{X}{n x q matrix of covariates}

\item{X_test}{matrix of test covariates}

\item{rho}{scaling parameter}
}
\value{
Gram matrix for specified kernel
}
\description{
Kernel functions
}

