% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.multiKernel}
\alias{predict.multiKernel}
\title{Prediction method for multivariate kernel-machine regression}
\usage{
predict.multiKernel(object, covariate, confounder = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{multiKernel}, such as the output of the 
function \code{\link{fitMultiKernel}}.}

\item{covariate}{Matrix of covariate variables, which are included in the 
kernel.}

\item{confounder}{Matrix or data.frame of confounder variables, which are not
included in the kernel.}

\item{...}{Extra parameters to be passed to the kernel function.}
}
\description{
This method performs prediction for new data, based on the output of a fitted
multivariate kernel-machine regression
}

