% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiKernel.R
\name{selectMultiKernel}
\alias{selectMultiKernel}
\title{Selection of tuning parameter for multivariate kernel regression}
\usage{
selectMultiKernel(response, covariate, confounder = NULL,
  kernel = c("linear", "quadratic", "gaussian"), tau_seq, K = 5,
  pure = FALSE)
}
\arguments{
\item{response}{matrix of response variables}

\item{covariate}{matrix of covariate variables, which are included in the
kernel.}

\item{confounder}{matrix or data.frame of confounder variables, which are not
included in the kernel.}

\item{kernel}{Type of kernel to use.}

\item{tau_seq}{Sequence of tuning parameters.}

\item{K}{number of folds for cross-validation.}

\item{pure}{Logical. Use the pure R version?}
}
\value{
Returns a list of kernel predictors, indexed by the different values
  of tau.
}
\description{
This function performs cross-validation for multivariate kernel regression
and selects the optimal tuning parameter among a user-specified collection
}

